requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.1.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
